backend/requirements.txt
Flask
flask-cors
Pillow
tensorflow
numpy
